+++ botsettings.json
@@ -1,6 +1,6 @@
 {
-    "token": "<REDACTED>",
-    "prefix": "dr!",
+    "token": "NDg1MTg4NjM5Nzk4NTkxNTA4.Dm9Zew.08-YVJOztorHezJlDDYlUqew-Mk",
+    "prefix": "+",
     "yt_api_key": "<REDACTED>",
     "idioticapi": "<REDACTED>",
     "dbltoken": "<REDACTED>"
