+++ commands/addCommand.txt
@@ -1,0 +1,9 @@
+module.exports.run = async (bot, message, args) => {
+
+}
+
+
+module.exports.help = {
+    name: "ban",
+    description:"Invoke the great power of the BAN HAMMER on someone."
+}
