+++ events/addEvent.txt
@@ -1,0 +1,6 @@
+exports.run = (bot, paramter) => 
+
+{
+
+    
+}
