+++ events/message.js
@@ -1,0 +1,25 @@
+const botSettings = require("../botsettings.json");
+const prefix = botSettings.prefix;
+exports.run = async (bot, message) => {
+    if(message.content === botSettings.token){
+        message.delete(0);
+    }
+if(message.author.bot) return;
+if(message.channel.type === "dm") return;
+
+let messageArray = message.content.split(/\s+/g);
+let command = messageArray[0];
+let args = messageArray.slice(1);
+
+if (message.channel.id === "450830883440558091") {
+  console.log("New suggestion in " + message.channel.name);
+  message.react("👍").then(message.react("👎"));
+}
+
+if(!command.startsWith(prefix)) return;
+
+let cmd = bot.commands.get(command.slice(prefix.length))
+if(cmd) cmd.run(bot, message, args);
+
+
+}
